+++ release/create.go
@@ -3,7 +3,6 @@ package main
 import (
 	"context"
 	"dagger/release/internal/dagger"
-	"dagger/release/util"
 	"fmt"
 	"strings"
 
@@ -12,59 +11,9 @@ import (
 
 // This file hosts general publish functions used by project specific publish methods.
 
-const (
-	imageOras = "ghcr.io/oras-project/oras:v1.2.3"
-	// A dagger module does exist for glab, but doesn't support custom base images,includes the deprecated gitlab release CLI, and doesn't have first-class support for release assets.
-	// https://github.com/vbehar/daggerverse/tree/838974e23bf2afb192850103c6d76fe620f31afd/gitlab-cli
-	imageGlabCLI = "registry.gitlab.com/gitlab-org/cli:latest"
-)
-
-// Publish additional tags to a remote OCI artifact.
-func (r *Release) AddTags(ctx context.Context,
-	// Existing OCI reference
-	ref string,
-	// Additional tags
-	tags []string,
-) (string, error) {
-	return r.orasCtr().
-		WithExec(append([]string{"oras", "tag", ref}, tags...)).
-		Stdout(ctx)
-}
-
-// Generate extra tags based on the provided target tag.
-//
-// Ex: Given the patch release 'v1.2.3', with an existing 'v1.3.0' release, it returns 'v1.2'.
-// Ex: Given the patch release 'v1.2.3', which is the latest and greatest, it returns 'v1', 'v1.2', 'latest'.
-//
-// Notice: current issue with SSH AUTH SOCK: https://docs.dagger.io/api/remote-repositories/#multiple-ssh-keys-may-cause-ssh-forwarding-to-fail
-func (r *Release) ExtraTags(ctx context.Context,
-	// OCI repository, e.g. localhost:5000/helloworld
-	ref string,
-	// target version
-	version string,
-) ([]string, error) {
-	out, err := r.orasCtr().
-		WithExec([]string{"oras", "repo", "tags", ref}).
-		Stdout(ctx)
-	if err != nil {
-		return nil, fmt.Errorf("retrieving existing repository tags: %w", err)
-	}
-	existing := strings.Fields(out)
-
-	return util.ExtraTags(version, existing)
-}
-
-// orasCtr returns a container with an oras executable, with mounted registry credentials.
-func (r *Release) orasCtr() *dagger.Container {
-	oras := dag.Container().
-		From(imageOras).
-		File("/bin/oras")
-
-	return dag.Wolfi().
-		Container().
-		WithMountedFile("/bin/oras", oras).
-		WithMountedSecret("/root/.docker/config.json", r.RegistryConfig.Secret())
-}
+// A dagger module does exist for glab, but doesn't support custom base images,includes the deprecated gitlab release CLI, and doesn't have first-class support for release assets.
+// https://github.com/vbehar/daggerverse/tree/838974e23bf2afb192850103c6d76fe620f31afd/gitlab-cli
+const imageGlabCLI = "registry.gitlab.com/gitlab-org/cli:latest"
 
 // Create a release in GitHub.
 func (r *Release) CreateGithub(ctx context.Context,
@@ -72,11 +21,11 @@ func (r *Release) CreateGithub(ctx context.Context,
 	repo string,
 	// gitlab personal access token
 	token *dagger.Secret,
-	// Release version
-	version string,
+	// tag to create release with
+	tag string,
 	// Release notes file
 	notes *dagger.File,
-	// Release title. Default: version
+	// Release title. Default: tag
 	// +optional
 	title string,
 	// Release assets
@@ -84,18 +33,14 @@ func (r *Release) CreateGithub(ctx context.Context,
 	assets []*dagger.File,
 ) (string, error) {
 
-	if title == "" {
-		title = version
-	}
-
 	err := dag.Gh(
 		dagger.GhOpts{
 			Token:  token,
 			Repo:   repo,
-			Source: r.gitRefAsDir(r.GitRef),
+			Source: r.GitRef.Tree(),
 		}).
 		Release().
-		Create(ctx, version, title,
+		Create(ctx, tag, title,
 			dagger.GhReleaseCreateOpts{
 				NotesFile: notes,
 				Files:     assets,
@@ -116,17 +61,18 @@ func (r *Release) CreateGitlab(ctx context.Context,
 	project string,
 	// GitLab personal access token
 	token *dagger.Secret,
-	// Release version
-	version string,
+	// Release tag
+	tag string,
 	// Release notes file
 	notes *dagger.File,
-	// Release title. Default: version
+	// Release title. Default: tag
 	// +optional
 	title string,
 	// Release assets
 	// +optional
 	assets []*dagger.File,
 ) (string, error) {
+
 	notesFileName, err := notes.Name(ctx)
 	if err != nil {
 		return "", err
@@ -134,7 +80,7 @@ func (r *Release) CreateGitlab(ctx context.Context,
 
 	hostRepo := fmt.Sprintf("%s/%s", host, project)
 	if title == "" {
-		title = version
+		title = tag
 	}
 
 	_, err = dag.Container().
@@ -143,8 +89,8 @@ func (r *Release) CreateGitlab(ctx context.Context,
 		WithSecretVariable("GITLAB_TOKEN", token).
 		WithEnvVariable("GITLAB_HOST", host).
 		WithExec([]string{"glab", "release", "create",
-			"-R", project, // repository
-			version, // tag
+			"-R=" + project, // repository
+			tag,             // tag
 			"--name=" + title,
 			"--notes-file=" + notesFileName, // description
 		}).
@@ -154,9 +100,9 @@ func (r *Release) CreateGitlab(ctx context.Context,
 	}
 
 	if len(assets) > 0 {
-		_, err := r.gitlabUploadAssets(ctx, host, project, token, version, assets)
+		_, err := r.gitlabUploadAssets(ctx, host, project, token, tag, assets)
 		if err != nil {
-			return "", fmt.Errorf("uploading release assets for '%s' release '%s': %w", hostRepo, version, err)
+			return "", fmt.Errorf("uploading release assets for '%s' release '%s': %w", hostRepo, tag, err)
 		}
 	}
 
@@ -171,8 +117,8 @@ func (r *Release) gitlabUploadAssets(ctx context.Context,
 	project string,
 	// GitLab personal access token
 	token *dagger.Secret,
-	// Release version
-	version string,
+	// Release tag
+	tag string,
 	// Release assets
 	assets []*dagger.File,
 ) (string, error) {
@@ -193,7 +139,7 @@ func (r *Release) gitlabUploadAssets(ctx context.Context,
 				WithEnvVariable("GITLAB_HOST", host).
 				WithExec([]string{"glab", "release", "upload",
 					"-R", project,
-					"v" + version,
+					tag,
 					assetName},
 				).
 				Stdout(ctx)