+++ docker/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.8] - 2025-11-20
+
+### 🐛 Bug Fixes
+
+- Bug in publish with build call [33a193e](https://github.com/act3-ai/dagger/commit/33a193ed062c3ec0020f569378feba209d844d22) by @pspurlock
+
+
+## [0.1.7] - 2025-11-20
+
+### 🐛 Bug Fixes
+
+- Updating dagger engine to v0.19.6 [8495191](https://github.com/act3-ai/dagger/commit/84951910d7bf3679f1789e93fa3b313ad7032694) by @pspurlock
+
+
+## [0.1.6] - 2025-11-18
+
+### 🐛 Bug Fixes
+
+- Upgrade dagger engine to v0.19.5 [50890a8](https://github.com/act3-ai/dagger/commit/50890a89d660ee551deefed7b79e84d32106bfa3) by @pspurlock
+
+- Add option for alternate dockerfile location and set target default to none [a062625](https://github.com/act3-ai/dagger/commit/a062625b6fda691c5b2064773f42eac083cf4c3a) by @pspurlock
+
+
+## [0.1.5] - 2025-11-06
+
+### 🐛 Bug Fixes
+
+- Add cliff.toml [521456d](https://github.com/act3-ai/dagger/commit/521456d6d2765b41e8a8667d22a2304f13c0c6c8) by @pspurlock
+
+
 ## [docker/v0.1.4] - 2025-10-02
 
 ### 🐛 Bug Fixes