+++ renovate/CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.11] - 2025-11-20
+
+### 🐛 Bug Fixes
+
+- Updating dagger engine to v0.19.6 [c18c6da](https://github.com/act3-ai/dagger/commit/c18c6da7c94228b172f24812ba2e89e4f7704c92) by @pspurlock
+
+
+## [0.1.10] - 2025-11-06
+
+### 🐛 Bug Fixes
+
+- Add cliff.toml [521456d](https://github.com/act3-ai/dagger/commit/521456d6d2765b41e8a8667d22a2304f13c0c6c8) by @pspurlock
+
+
 ## [renovate/v0.1.9] - 2025-09-19
 
 ### 🐛 Bug Fixes